+++ R/count_linkage.R
@@ -8,19 +8,23 @@
 #'  If NULL, only total counts of linkages for linkages in the cluster across all subjects is given.
 #' @param linkage a stored linkage from the domino object.
 #'  Can compare any of 'tfs', 'rec', 'incoming_lig', 'tfs_rec', or 'rec_lig'
-#' @param subject_names a vector of subject_names from the linkage_summary.
-#'   NOTE: all subject_names in the linkage summary are included in counting.
+#' @param subject_names a vector of subject_names from the linkage_summary to restrict counting to.
+#'   If NULL, all subjects in `link_summary` are included.
 #' @return A data frame with columns for the unique linkage features and the counts of how many times the linkage
-#'  occured across the compared domino results. If group.by is used, counts of the linkages are also provided as
+#'  occurred across the compared domino results. If group.by is used, counts of the linkages are also provided as
 #'  columns named by the unique values of the group.by variable.
 #' @export
+#' @seealso [summarize_linkages()] to create [linkage_summary()] objects, [test_differential_linkages()] to compare links across groups
 #' @family differentials
 #' @examples
 #' data(LinkageSummary)
 #' count_linkage(
 #'   link_summary = LinkageSummary$linkage_sum_tiny, cluster = "C1", 
 #'   group.by = "group", linkage = "rec")
-#' 
+#' count_linkage(
+#'   link_summary = LinkageSummary$linkage_sum_tiny, cluster = "C1", 
+#'   group.by = "group", linkage = "tfs_rec", subject_names = c("P1", "P2", "P4", "P5"))
+
 count_linkage <- function(link_summary, cluster, group.by = NULL, linkage = "rec_lig", subject_names = NULL) {
     check_arg(link_summary, allow_class = "linkage_summary", allow_len = 1)
     check_arg(cluster, allow_class = "character", allow_len = 1)
@@ -29,9 +33,7 @@ count_linkage <- function(link_summary, cluster, group.by = NULL, linkage = "rec
         allow_values = c("tfs", "rec", "incoming_lig", "tfs_rec", "rec_lig"))
     check_arg(subject_names, allow_class = c("factor", "character", "NULL"))
 
-    if (is.null(subject_names)) {
-        subject_names <- link_summary@subject_names
-    }
+    link_summary <- filter_by_subject_names(link_summary, subject_names)
     all_int_ls <- lapply(link_summary@subject_linkages, FUN = function(x) {
         return(x[[cluster]][[linkage]])
     })